+++ jars/orm/src/main/java/it/govpay/bd/model/PagamentoPortale.java
@@ -20,6 +20,7 @@
 package it.govpay.bd.model;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.text.MessageFormat;
 import java.util.ArrayList;
 import java.util.Date;
@@ -118,7 +119,7 @@ public class PagamentoPortale extends BasicModel {
 	private int tipo;
 	private boolean ack;
 	
-	private Integer severita;
+	private BigInteger severita;
 	
 	public String getIdSessione() {
 		return this.idSessione;
@@ -339,10 +340,10 @@ public class PagamentoPortale extends BasicModel {
 	public void setIdApplicazione(Long idApplicazione) {
 		this.idApplicazione = idApplicazione;
 	}
-	public Integer getSeverita() {
+	public BigInteger getSeverita() {
 		return severita;
 	}
-	public void setSeverita(Integer severita) {
+	public void setSeverita(BigInteger severita) {
 		this.severita = severita;
 	}
